+++ IntlFcStoneCodeChallenge/IntlFcStoneCodeChallenge/Interfaces/IContactRepository.cs
@@ -1,10 +1,11 @@
 ﻿using IntlFcStoneCodeChallenge.Models;
+using System;
 
 namespace IntlFcStoneCodeChallenge.Interfaces
 {
     interface IContactRepository : IBaseRepository<Contact>
     {
-        bool GetByPhone(int phone, out Contact contact);
+        bool GetByPhone(UInt64 phone, out Contact contact);
         bool GetByEmail(string email, out Contact contact);
     }
 }
